Improved glow logic for items slightly
Enough for you to notice the update/change in game, you will like it better, trust us.  Also got rid of all unnecessary files on the repo so nobody gets confused, would not want that happening.  Either grab the .zip and save it to your flash drive  and then unzip the .exe there, it is how we do it, and how i recommend you do it as well.  But, for those that are stubborn, here is the .exe if you cant afford the extra extracting step. 
As young wattson would say, "PRED FOR ALL!!"  Enjoy.